added checking owner

+++ controllers/contacts.js
@@ -5,7 +5,7 @@ const { ctrlWrapper } = require("../helpers");
 
 const getAll = async (req, res) => {
   const { _id: owner } = req.user;
-  const { page = 1, limit = 20, favorite=[true, false] } = req.query;
+  const { page = 1, limit = 20, favorite = [true, false] } = req.query;
   const skip = (page - 1) * limit;
   const resultContacts = await Contact.find(
     { owner, favorite },
@@ -21,9 +21,12 @@ const getAll = async (req, res) => {
 
 const getById = async (req, res) => {
   const { contactId } = req.params;
-  console.log(contactId);
+  const { _id: owner } = req.user;
   const resultContact = await Contact.findById(contactId);
   if (!resultContact) {
+    throw HttpError(404, "Not found");
+  }
+  if (!resultContact.owner.equals(owner)) {
     throw HttpError(404, "Not found");
   }
   res.json(resultContact);
@@ -36,33 +39,51 @@ const add = async (req, res) => {
 };
 
 const deleteById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const deletedResult = await Contact.findByIdAndDelete(contactId);
-  if (!deletedResult) {
+  const resultContact = await Contact.findById(contactId);
+  if (!resultContact) {
     throw HttpError(404, "Not found");
   }
+  if (!resultContact.owner.equals(owner)) {
+    throw HttpError(404, "Not found");
+  }
+
+  await Contact.findByIdAndDelete(contactId);
   res.json({ message: "contact deleted" });
 };
 
 const updateById = async (req, res) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
+  const resultContact = await Contact.findById(contactId);
+  if (!resultContact) {
+    throw HttpError(404, "Not found");
+  }
+  if (!resultContact.owner.equals(owner)) {
+    throw HttpError(404, "Not found");
+  }
+
   const updatedResult = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
-  if (!updatedResult) {
-    throw HttpError(404, "Not found");
-  }
   res.json(updatedResult);
 };
 
 const updateStatusContact = async (req, res) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
+  const resultContact = await Contact.findById(contactId);
+  if (!resultContact) {
+    throw HttpError(404, "Not found");
+  }
+  if (!resultContact.owner.equals(owner)) {
+    throw HttpError(404, "Not found");
+  }
+
   const updatedResult = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
-  if (!updatedResult) {
-    throw HttpError(404, "Not found");
-  }
   res.json(updatedResult);
 };
